refactor(firestore): extract atomic write coordinator
